Move to hosting our own classes
