+++ lib/hanami/minitest/generators/part/part_base_test.erb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class <%= camelized_app_name %>::Views::PartTest < Minitest::Test
+class <%= camelized_app_name %>::Views::PartTest < Hanami::Minitest::Test
   def setup
     @value = Object.new
     @subject = <%= camelized_app_name %>::Views::Part.new(value: @value)
