+++ lib/hanami/minitest/generators/part/part_slice_test.erb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class <%= camelized_slice_name %>::Views::Parts::<%= camelized_name %>Test < Minitest::Test
+class <%= camelized_slice_name %>::Views::Parts::<%= camelized_name %>Test < Hanami::Minitest::Test
   def setup
     @value = Object.new
     @subject = <%= camelized_slice_name %>::Views::Parts::<%= camelized_name %>.new(value: @value)
