+++ lib/hanami/minitest/generators/slice/action_test.erb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class <%= camelized_slice_name %>::ActionTest < Minitest::Test
+class <%= camelized_slice_name %>::ActionTest < Hanami::Minitest::Test
   def test_pending
     skip "Add tests for <%= slice %> actions"
   end
