+++ lib/hanami/minitest/generators/support_requests.rb
@@ -1,12 +1,5 @@
 # frozen_string_literal: true
 
-require "rack/test"
-
-class RequestTest < Minitest::Test
-  include Rack::Test::Methods
-
-  # Defines the app for Rack::Test requests.
-  def app
-    Hanami.app
-  end
+class Hanami::Minitest::RequestTest
+  # Add custom request test helpers here.
 end
