+++ lib/hanami/minitest/test.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require "minitest/test"
+
+module Hanami
+  module Minitest
+    # Base test class for Hanami applications.
+    #
+    # @since 2.0.0
+    # @api public
+    class Test < ::Minitest::Test
+    end
+  end
+end
